le_streamer_client: fix warning

+++ example/le_streamer_client.c
@@ -138,7 +138,7 @@ static void test_track_data(le_streamer_client_connection_t * context, int bytes
     if (time_passed < REPORT_INTERVAL_MS) return;
     // print speed
     uint32_t bytes_per_second = context->test_data_sent * 1000 / time_passed;
-    printf("%c: %"PRIu32" bytes -> %u.%03u kB/s\n", context->name, context->test_data_sent, bytes_per_second / 1000, bytes_per_second % 1000);
+    printf("%c: %"PRIu32" bytes -> %u.%03u kB/s\n", context->name, context->test_data_sent, (int) (bytes_per_second / 1000), (int)(bytes_per_second % 1000));
 
     // restart
     context->test_data_start = now;
